+++ src/NuGet.Updater/Entities/UpdaterParameters.cs
@@ -10,14 +10,14 @@ namespace NuGet.Updater.Entities
 		public string SolutionRoot { get; set; }
 
 		/// <summary>
-		/// Gets or sets a list of private feed URLs and access tokens to get packages from.
+		/// Gets or sets a list of source to get packages from.
 		/// </summary>
-		public Dictionary<string, string> PrivateFeeds { get; set; }
+		public ICollection<IUpdaterSource> Sources { get; set; }
 
 		/// <summary>
 		/// Gets or sets the versions to update to (stable, dev, beta, etc.), in order of priority.
 		/// </summary>
-		public IEnumerable<string> TargetVersions { get; set; } = new[] { "stable" };
+		public ICollection<string> TargetVersions { get; set; } = new List<string> { "stable" };
 
 		/// <summary>
 		/// Gets or sets a value indicating whether the version should exactly match the target version.
@@ -42,12 +42,12 @@ namespace NuGet.Updater.Entities
 		/// <summary>
 		/// Gets or sets a list of packages to ignore.
 		/// </summary>
-		public IEnumerable<string> PackagesToIgnore { get; set; }
+		public ICollection<string> PackagesToIgnore { get; set; }
 
 		/// <summary>
 		/// Gets or sets a list of packages to update; all packages found will be updated if nothing is specified.
 		/// </summary>
-		public IEnumerable<string> PackagesToUpdate { get; set; }
+		public ICollection<string> PackagesToUpdate { get; set; }
 
 		/// <summary>
 		/// Gets or sets the name of the author of the packages to update; used with NuGet.org; packages from private feeds are assumed to be required.